Move setters macro to core crate
