+++ sdk/core/Cargo.toml
@@ -33,6 +33,7 @@ failure = "0.1"
 async-trait = "0.1.36"
 oauth2 = { version = "4.0.0-alpha.2" }
 reqwest = "*"
+paste = "1.0"
 
 [dev-dependencies]
 tokio = "0.2"
