+++ sdk/cosmos/Cargo.toml
@@ -27,7 +27,6 @@ serde_json = "1.0"
 url = "2.1"
 uuid = { version = "0.8", features = ["v4"] }
 failure = "0.1"
-paste = "1.0"
 
 [dev-dependencies]
 env_logger = "0.7"
